+++ constant/personality.go
@@ -1,0 +1,9 @@
+package constant
+
+var GetPersonalityTag = map[string]string{
+	"Extraversion":  "7,8,9",
+	"Neurotic":      "10,11,12",
+	"Agreeable":     "13,14,15",
+	"Conscientious": "16,17,18",
+	"Opennes":       "19,20,8",
+}