+++ CHANGELOG.md
@@ -77,3 +77,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### [1.0.8] - yusufw2429@gmail.com
 #### Added
 - Add X-API-Key for integration with ML (not using jwt)
+- Fixing bug
+- Update deployment
+- Add resty pkg utils
+- Integrate questionnaire with ML